requirements.txt
###rasa_nlu
gevent==1.2.1
Klein==17.2.0
boto3==1.4.4
typing==3.6.2
future==0.16.0
six==1.10.0
jsonschema==2.6.0
matplotlib==1.5.3
requests==2.14.2
tqdm==4.11.2
numpy==1.13.3
simplejson==3.11.1

### Sklearn
spacy==1.7.0
scikit-learn==0.19.1
sklearn-crfsuite==0.3.5

###rasa_core
apscheduler==3.3.1
fakeredis==0.8.2
graphviz==0.7.1
h5py==2.7.0
jsonpickle==0.9.4
keras==2.0.8
pandoc==1.0.0b2
redis==2.10.5
tensorflow==1.8.0rc1
networkx==1.11
fbmessenger==5.0.0
ConfigArgParse==0.12.0
pykwalify==1.6.0
coloredlogs==7.3
ruamel.yaml==0.15.34
flask==0.12
rasa_nlu==0.12.3

###additional
pypandoc
slackclient
rasa_core==0.9.0a2
git+https://github.com/apixu/apixu-python.git
